services/visit_service/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
drf-spectacular
drf-yasg
requests>=2.31
pika==1.3.2
google-auth
djangorestframework-simplejwt
djangorestframework
requests
django-cors-headers==4.3.1
python-decouple
celery
django-celery-beat